refactor MarksList.tsx & some fixes
