+++ apps/web/src/views/Attestation/utils/index.ts
@@ -1,10 +1,6 @@
 import { AttestationResponse, Subject } from '@diary-spo/shared'
 
 export const processAttestationData = (data: AttestationResponse) => {
-  // if (!data?.students) {
-  //   return null
-  // }
-
   const student = data.students[0]
 
   const semesters: Record<string, Subject[]> = {}
