+++ apps/web/src/views/Marks/MarksByGroup/AverageMarkCell/helpers/__tests__/mocks.ts
