+++ apps/web/src/views/Marks/MarksByGroup/AverageMarkCell/helpers/index.ts
@@ -1,30 +1,6 @@
-import { Mark } from '@components'
-import { AbsenceType, AbsenceTypes, Grade, MarkKeys } from '@diary-spo/shared'
-import { CSSProperties, ReactNode } from 'preact/compat'
+import { Grade, MarkKeys } from '@diary-spo/shared'
+
 import { Nullable } from '@types'
-
-export const renderMarksOrAbsence = (
-  markValues: string[],
-  absenceType: AbsenceType
-): ReactNode => {
-  if (markValues.length && !absenceType) {
-    return markValues.map((mark, index) => (
-      <Mark key={index} mark={Grade[mark]} size='s' />
-    ))
-  }
-
-  if (absenceType) {
-    return <Mark size='s' mark={AbsenceTypes[absenceType]} />
-  }
-
-  return undefined
-}
-
-export const styles: CSSProperties = {
-  display: 'flex',
-  gap: 8,
-  marginLeft: 8
-}
 
 /**
  * Функция 'calculateAverageMark' высчитывает средний балл учащегося до двух знаков после запятой.
