+++ apps/web/src/views/Marks/MarksByGroup/MarksList/index.css
@@ -1,0 +1,9 @@
+.marksList {
+    display: flex;
+}
+
+.marksList__item {
+    display: flex;
+    gap: 8px;
+    margin-left: 8px;
+}
