+++ apps/web/src/views/Marks/MarksByGroup/MarksList/index.tsx
@@ -1,15 +1,15 @@
 import { SubjectMarksMap } from '@utils'
 import { FC } from 'preact/compat'
-import { renderMarksOrAbsence, styles } from './helpers'
+import { renderMarksOrAbsence } from './helpers'
 
 interface IMarksList {
   marks: SubjectMarksMap[string]
 }
 
 const MarksList: FC<IMarksList> = ({ marks }) => (
-  <div style={{ display: 'flex' }}>
+  <div className='marksList'>
     {marks.map(({ date, marks: markValues, absenceType }) => (
-      <div key={`${date}_${absenceType}`} style={styles}>
+      <div key={`${date}_${absenceType}`} className='marksList__item'>
         {renderMarksOrAbsence(markValues, absenceType)}
       </div>
     ))}
