+++ apps/web/src/views/Marks/Summary/index.tsx
@@ -3,15 +3,9 @@ import { VIOLET } from '@config'
 import { Icon20EducationOutline, Icon28BrainOutline } from '@vkontakte/icons'
 import { Group, Header, MiniInfoCell } from '@vkontakte/vkui'
 import { FC, Fragment } from 'preact/compat'
+import { ISummary } from './types.ts'
 
 import './index.css'
-import { Nullable } from '@types'
-
-interface ISummary {
-  totalNumberOfMarks: Nullable<number>
-  averageMark: Nullable<number>
-  markCounts: Nullable<Record<number, number>>
-}
 
 const Summary: FC<ISummary> = ({
   markCounts,
