+++ apps/web/src/views/Marks/Summary/types.ts
@@ -1,0 +1,7 @@
+import { Nullable } from '@types'
+
+export interface ISummary {
+  totalNumberOfMarks: Nullable<number>
+  averageMark: Nullable<number>
+  markCounts: Nullable<Record<number, number>>
+}
